Create README.md
Since this is the first JavaScript project being built from scratch, it’s important to remember the wise words from the Problem Solving lesson. For each step in this project, be sure to do the following

Plan or pseudocode your solution.
Write the code.
Test your code to make sure it works.